refactor GUI tests
