+++ src/test/java/org/cicirello/ibp/BottomPanelTests.java
@@ -23,170 +23,16 @@ package org.cicirello.ibp;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-import java.awt.BorderLayout;
 import java.awt.Component;
 import java.util.ArrayList;
 import javax.swing.JButton;
 import javax.swing.JComboBox;
 import javax.swing.JLabel;
 import javax.swing.JPanel;
-import javax.swing.JTextArea;
-import javax.swing.JTextField;
 import org.junit.jupiter.api.*;
 
-/**
- * JUnit tests for all of the GUI classes of the Interactive Bin Packing Application. These test
- * cases automates testing nearly all of the GUI, simulating button presses and menu item clicks,
- * etc via Swing's doClick() method. The only GUI elements not tested via JUnit, which will require
- * manual testing are cases where dialog boxes are displayed to inform the user of something, as
- * well as hyperlinks within the Tutorial and Help dialogs.
- */
-public class GUITestCases {
-
-  @Test
-  public void testTopPanel() {
-    int[] sizes = {7, 2};
-    CallBack cb =
-        new CallBack() {
-          @Override
-          public void call() {}
-        };
-    Floor floor = new Floor(sizes);
-    ApplicationState state = new ApplicationState(1, floor, cb, cb, cb);
-    TopPanel top = new TopPanel(state);
-    BorderLayout layout = (BorderLayout) top.getLayout();
-    JTextArea text = (JTextArea) layout.getLayoutComponent(BorderLayout.CENTER);
-    String expected = "A(7), B(2)";
-    assertEquals(expected, text.getText());
-    expected = "Not yet in any bins: ";
-    JLabel label = (JLabel) layout.getLayoutComponent(BorderLayout.NORTH);
-    assertEquals(expected, label.getText());
-    state.getFloor().add(new Item("C", 15));
-    top.refresh();
-    expected = "A(7), B(2), C(15)";
-    assertEquals(expected, text.getText());
-    expected = "Not yet in any bins: ";
-    assertEquals(expected, label.getText());
-  }
-
-  @Test
-  public void testWestPanel() {
-    int[] sizes = {7, 2};
-    int[] sizes2 = {5, 8};
-    CallBack cb =
-        new CallBack() {
-          @Override
-          public void call() {}
-        };
-    for (int numBins = 1; numBins <= 3; numBins++) {
-      Floor floor = new Floor(sizes);
-      ApplicationState state = new ApplicationState(numBins, floor, cb, cb, cb);
-      WestPanel west = new WestPanel(numBins, state);
-      Component[] c = west.getComponents();
-      assertEquals(numBins, c.length);
-      for (int i = 0; i < c.length; i++) {
-        JPanel row = (JPanel) c[i];
-        assertEquals("Capacity: 100", ((JLabel) row.getComponent(0)).getText().trim());
-        assertEquals("Used: 0", ((JLabel) row.getComponent(1)).getText().trim());
-      }
-      state.getBins().get(0).add(new Item("C", 10));
-      west.refresh();
-      for (int i = 0; i < c.length; i++) {
-        JPanel row = (JPanel) c[i];
-        assertEquals("Capacity: 100", ((JLabel) row.getComponent(0)).getText().trim());
-        if (i != 0) {
-          assertEquals("Used: 0", ((JLabel) row.getComponent(1)).getText().trim());
-        } else {
-          assertEquals("Used: 10", ((JLabel) row.getComponent(1)).getText().trim());
-        }
-      }
-      state.getBins().get(0).add(new Item("D", 90));
-      west.refresh();
-      for (int i = 0; i < c.length; i++) {
-        JPanel row = (JPanel) c[i];
-        assertEquals("Capacity: 100", ((JLabel) row.getComponent(0)).getText().trim());
-        if (i != 0) {
-          assertEquals("Used: 0", ((JLabel) row.getComponent(1)).getText().trim());
-        } else {
-          assertEquals("Used: 100", ((JLabel) row.getComponent(1)).getText().trim());
-        }
-      }
-    }
-    Floor floor = new Floor(sizes);
-    int numBins = 4;
-    ApplicationState state = new ApplicationState(numBins, floor, cb, cb, cb);
-    ArrayList<Bin> bins = state.getBins();
-    bins.get(0).add(new Item("C", 5));
-    bins.get(1).add(new Item("D", 10));
-    bins.get(2).add(new Item("E", 100));
-    WestPanel west = new WestPanel(numBins, state);
-    Component[] c = west.getComponents();
-    assertEquals(numBins, c.length);
-    for (int i = 0; i < c.length; i++) {
-      JPanel row = (JPanel) c[i];
-      assertEquals("Capacity: 100", ((JLabel) row.getComponent(0)).getText().trim());
-      switch (i) {
-        case 0:
-          assertEquals("Used: 5", ((JLabel) row.getComponent(1)).getText().trim());
-          break;
-        case 1:
-          assertEquals("Used: 10", ((JLabel) row.getComponent(1)).getText().trim());
-          break;
-        case 2:
-          assertEquals("Used: 100", ((JLabel) row.getComponent(1)).getText().trim());
-          break;
-        case 3:
-          assertEquals("Used: 0", ((JLabel) row.getComponent(1)).getText().trim());
-          break;
-      }
-    }
-  }
-
-  @Test
-  public void testCenterPanel() {
-    int[] sizes = {7, 2};
-    CallBack cb =
-        new CallBack() {
-          @Override
-          public void call() {}
-        };
-    int[] binCounts = {1, 2, 9, 10};
-    for (int b = 0; b < binCounts.length; b++) {
-      int numBins = binCounts[b];
-      Floor floor = new Floor(sizes);
-      ApplicationState state = new ApplicationState(numBins, floor, cb, cb, cb);
-      CenterPanel mid = new CenterPanel(numBins, state);
-      Component[] c = mid.getComponents();
-      assertEquals(numBins, c.length);
-      for (int i = 0; i < c.length; i++) {
-        JPanel row = (JPanel) c[i];
-        assertEquals("Bin " + (i + 1) + ":", ((JLabel) row.getComponent(0)).getText().trim());
-        assertEquals("empty", ((JTextField) row.getComponent(1)).getText().trim());
-      }
-      state.getBins().get(numBins - 1).add(new Item("C", 10));
-      mid.refresh();
-      for (int i = 0; i < c.length; i++) {
-        JPanel row = (JPanel) c[i];
-        assertEquals("Bin " + (i + 1) + ":", ((JLabel) row.getComponent(0)).getText().trim());
-        if (i != numBins - 1) {
-          assertEquals("empty", ((JTextField) row.getComponent(1)).getText().trim());
-        } else {
-          assertEquals("C(10)", ((JTextField) row.getComponent(1)).getText().trim());
-        }
-      }
-      state.getBins().get(numBins - 1).add(new Item("D", 13));
-      mid.refresh();
-      for (int i = 0; i < c.length; i++) {
-        JPanel row = (JPanel) c[i];
-        assertEquals("Bin " + (i + 1) + ":", ((JLabel) row.getComponent(0)).getText().trim());
-        if (i != numBins - 1) {
-          assertEquals("empty", ((JTextField) row.getComponent(1)).getText().trim());
-        } else {
-          assertEquals("C(10), D(13)", ((JTextField) row.getComponent(1)).getText().trim());
-        }
-      }
-    }
-  }
+/** JUnit tests for the bottom panel. */
+public class BottomPanelTests {
 
   @Test
   public void testBottomPanel() {
